+++ apps/backend/src/modules/tasks/task.controller.ts
@@ -9,44 +9,54 @@ import { type Logger } from "~/libs/modules/logger/logger.js";
 import { type UserDto } from "~/modules/users/users.js";
 
 import { TasksApiPath } from "./libs/enums/enums.js";
+import { checkAccessToTask } from "./libs/hooks/hooks.js";
+import {
+	type TaskUpdateParametersDto,
+	type TaskUpdateRequestDto,
+} from "./libs/types/types.js";
+import { taskUpdateValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
 import { type TaskService } from "./task.service.js";
 
-/*** @swagger
+/**
+ * @swagger
+ * tags:
+ *   - name: tasks
+ *     description: Endpoints related to tasks
  * components:
- *    schemas:
- *      Task:
- *        type: object
- *        properties:
- *          id:
- *            type: number
- *            format: number
- *            minimum: 1
- *          category:
- *            type: string
- *          categoryId:
- *            type: number
- *            format: number
- *            minimum: 1
- *          description:
- *            type: string
- *          dueDate:
- *            type: string
- *          label:
- *            type: string
- *          status:
- *            type: string
- *          userId:
- *            type: number
- *            format: number
- *            minimum: 1
- *          createdAt:
- *            type: string
- *            format: date-time
- *          updatedAt:
- *            type: string
- *            format: date-time
+ *   schemas:
+ *     TaskStatusEnum:
+ *       type: string
+ *       enum: [Current, Completed, Skipped]
+ *     TaskDto:
+ *       type: object
+ *       properties:
+ *         id:
+ *           type: integer
+ *           example: 1
+ *         categoryId:
+ *           type: integer
+ *           example: 1
+ *         userId:
+ *           type: integer
+ *           example: 1
+ *         category:
+ *           type: string
+ *         description:
+ *           type: string
+ *         dueDate:
+ *           type: string
+ *           format: date-time
+ *         label:
+ *           type: string
+ *         status:
+ *           $ref: '#/components/schemas/TaskStatusEnum'
+ *         createdAt:
+ *           type: string
+ *           format: date-time
+ *         updatedAt:
+ *           type: string
+ *           format: date-time
  */
-
 class TaskController extends BaseController {
 	private taskService: TaskService;
 
@@ -65,13 +75,78 @@ class TaskController extends BaseController {
 			method: "GET",
 			path: TasksApiPath.CURRENT,
 		});
+
+		this.addRoute({
+			handler: (options) =>
+				this.findPastByUserId(
+					options as APIHandlerOptions<{
+						user: UserDto;
+					}>,
+				),
+			method: "GET",
+			path: TasksApiPath.PAST,
+		});
+
+		this.addRoute({
+			handler: (options) =>
+				this.update(
+					options as APIHandlerOptions<{
+						body: TaskUpdateRequestDto;
+						params: TaskUpdateParametersDto;
+					}>,
+				),
+			method: "PATCH",
+			path: TasksApiPath.$ID,
+			preHandlers: [checkAccessToTask(taskService)],
+			validation: {
+				body: taskUpdateValidationSchema,
+			},
+		});
 	}
 
 	/**
 	 * @swagger
 	 * /tasks/current:
+	 *   get:
+	 *     tags: [tasks]
+	 *     summary: Get user current tasks
+	 *     security:
+	 *       - bearerAuth: []
+	 *     responses:
+	 *       200:
+	 *         description: Successful operation
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               type: array
+	 *               items:
+	 *                 $ref: "#/components/schemas/TaskDto"
+	 *       401:
+	 *         description: Unauthorized
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: "#/components/schemas/CommonErrorResponse"
+	 */
+	private async findCurrentByUserId(
+		options: APIHandlerOptions<{
+			user: UserDto;
+		}>,
+	): Promise<APIHandlerResponse> {
+		const { user } = options;
+
+		return {
+			payload: await this.taskService.findCurrentByUserId(user.id),
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /tasks/past:
 	 *    get:
-	 *      description: Returns an array of current users tasks
+	 *      tags: [tasks]
+	 *      summary: Get user past tasks
 	 *      security:
 	 *        - bearerAuth: []
 	 *      responses:
@@ -82,10 +157,15 @@ class TaskController extends BaseController {
 	 *              schema:
 	 *                type: array
 	 *                items:
-	 *                  $ref: "#/components/schemas/Task"
+	 *                  $ref: "#/components/schemas/TaskDto"
+	 *        401:
+	 *          description: Unauthorized
+	 *          content:
+	 *            application/json:
+	 *              schema:
+	 *                $ref: "#/components/schemas/CommonErrorResponse"
 	 */
-
-	private async findCurrentByUserId(
+	private async findPastByUserId(
 		options: APIHandlerOptions<{
 			user: UserDto;
 		}>,
@@ -93,7 +173,70 @@ class TaskController extends BaseController {
 		const { user } = options;
 
 		return {
-			payload: await this.taskService.findCurrentByUserId(user.id),
+			payload: await this.taskService.findPastByUserId(user.id),
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /tasks/{id}:
+	 *   patch:
+	 *     tags: [tasks]
+	 *     summary: Update task status by id
+	 *     security:
+	 *       - bearerAuth: []
+	 *     parameters:
+	 *       - name: id
+	 *         in: path
+	 *         required: true
+	 *         description: Task id
+	 *         schema:
+	 *           type: integer
+	 *           example: 1
+	 *     requestBody:
+	 *       required: true
+	 *       content:
+	 *         application/json:
+	 *           schema:
+	 *             type: object
+	 *             properties:
+	 *               status:
+	 *                 $ref: "#/components/schemas/TaskStatusEnum"
+	 *     responses:
+	 *       200:
+	 *         description: Successful operation
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: "#/components/schemas/TaskDto"
+	 *       401:
+	 *         description: Unauthorized
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: "#/components/schemas/CommonErrorResponse"
+	 *       403:
+	 *         description: Forbidden to update other user's tasks
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: "#/components/schemas/CommonErrorResponse"
+	 *       422:
+	 *         description: Validation error
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               $ref: "#/components/schemas/ValidationErrorResponse"
+	 */
+	private async update(
+		options: APIHandlerOptions<{
+			body: TaskUpdateRequestDto;
+			params: TaskUpdateParametersDto;
+		}>,
+	): Promise<APIHandlerResponse> {
+		return {
+			payload: await this.taskService.update(options.params.id, options.body),
 			status: HTTPCode.OK,
 		};
 	}