+++ apps/mobile/src/libs/enums/navigation/root-screen-name.enum.ts
@@ -1,6 +1,7 @@
 const RootScreenName = {
 	BOTTOM_TABS_NAVIGATOR: "BottomTabsNavigator",
 	CHAT: "Chat",
+	NOTIFICATION_QUESTIONS: "Notification questions",
 	ONBOARDING: "Onboarding",
 	QUIZ: "Quiz",
 	QUIZ_ENTRY: "Quiz Entry",